fix: #2495 time tracker play/pause issue
